feat: move notification in main modal content
