+++ packages/vechain-kit/src/components/common/ModalNotificationButton.tsx
@@ -20,7 +20,7 @@ export const ModalNotificationButton = ({
             _dark={{ _hover: { bg: 'whiteAlpha.100' } }}
             position="absolute"
             borderRadius={'50%'}
-            right="10px"
+            left="10px"
             top="10px"
             onClick={onClick}
             icon={
